+++ test/views/QuantumTP6m.test.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { BrowserRouter } from 'react-router-dom';
 import renderer from 'react-test-renderer';
-import QuantumTP6m from '../../src/quantum/TP6m';
+import QuantumTP6m from '../../src/views/Android/TP6m';
 
 it('renders correctly', () => {
   const tree = renderer
